pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support-agent"
version = "0.1.0"
description = "B2B SaaS support operations agent with retrieval, tools, and guardrails"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "openai>=1.0",
    "pyyaml>=6",
    "langchain-openai>=0.2",
    "langchain-core>=0.3",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "hypothesis>=6",
    "trajectly>=0.4",
]
eval = [
    "promptfoo>=0.1",
]

[project.scripts]
support-agent = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E402", "E501"]  # E402: import after path in evals; E501: long lines in data/tests
